Cleaned up WindowEventsListener class
